requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
itsdangerous==2.2.0
python-dotenv==1.0.1
databricks-sdk==0.40.0
pydantic==2.10.4
